Added all functionality
Added insertion sort into the program. Made it so you can easily toggle between the different sorting algorithms. Added the Text at the top of the screen to show what sorting algorithm is currently playing.

Fixes:
Fixed the red and green over top not appearing. The problem was I was redrawing overtop of it.

To do:
Make the menu be more appealing. Thinking of doing green bars, black background. Might possibly do red and blue over selection to show it is being selected