Added an average sort time calculator
Added a new python file that will be used in calculating the average times for the array to be sorted. Since the visual sorting algorithm is much slower this code will make it easier to determine the true speed of each algorithm. Currently only merge sort is implemented but it is very simple to add new algorithms.

The program currently has basic input handling